feat(download): Move download module to backend features
