+++ src/backend/features/backup/download-device-backup-zip.ts
@@ -3,7 +3,7 @@ import type { Device } from './types/Device';
 import { User } from '../../../apps/main/types';
 import { fetchFolder } from '../../../infra/drive-server/services/folder/services/fetch-folder';
 import { getCredentials } from '../../../apps/main/auth/get-credentials';
-import { downloadFolderAsZip } from '../../../apps/main/network/download/download-folder-as-zip';
+import { downloadFolderAsZip } from '../download/download-folder-as-zip';
 import { logger } from '@internxt/drive-desktop-core/build/backend';
 import { Result } from '../../../context/shared/domain/Result';
