+++ src/backend/features/download/get-required-file-metadata.ts
