+++ src/backend/features/download/stream/get-decrypted-stream.ts
