+++ src/backend/features/download/stream/is-empty-backup-file-without-file-id.ts
